Change the type of Account::m_balance from float to double

+++ Account.cpp
@@ -5,7 +5,7 @@ Account::Account()
     m_balance = 0.0F;
 }
 
-Account::Account(float initial_balance)
+Account::Account(double initial_balance)
 {
     m_balance = initial_balance;
 }
@@ -15,17 +15,17 @@ Account::Account(const Account& acc)
     m_balance = acc.m_balance;
 }
 
-float Account::get_balance() const
+double Account::get_balance() const
 {
     return m_balance;
 }
 
-void Account::deposit(float money)
+void Account::deposit(double money)
 {
     m_balance += money;
 }
 
-bool Account::withdraw(float money)
+bool Account::withdraw(double money)
 {
     if (money > m_balance)
         return false;
